Implement SpawnController class and create object in main.cpp

+++ src/Enemy.cpp
@@ -2,7 +2,7 @@
 #include <math.h>
 #include "Enemy.h"
 
-Enemy::Enemy(int id, float x, float y) : Entity(id, x, y)
+Enemy::Enemy(int id, double x, double y) : Entity(id, x, y)
 {
     std::cout << "Constructing enemy.. " << std::endl;
 }
